Fixed broken package names in template. Made api template primarily code, now much cleaner, fixed issue with config option being in incorrect namespace
