+++ packages-next/keystone/package.json
@@ -53,7 +53,8 @@
     "react-dom": "^17.0.1",
     "source-map-support": "^0.5.19",
     "typescript": "^4.1.5",
-    "uid-safe": "^2.1.5"
+    "uid-safe": "^2.1.5",
+    "apollo-server-micro":"^2.21.1"
   },
   "engines": {
     "node": ">=10.0.0"
