+++ packages-next/keystone/src/index.ts
@@ -1,3 +1,4 @@
 export { createSystem } from './lib/createSystem';
 export { createExpressServer } from './lib/createExpressServer';
 export { initConfig } from './lib/initConfig';
+export { createApolloServerMicro } from './lib/createApolloServerMicro';
